session added

+++ delete_task.php
@@ -2,11 +2,16 @@
 include 'config.php';
 
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+    session_start();
+    if (!isset($_SESSION['user_id'])) {
+        header("Location: ./login/log_in.php");
+        exit();
+    }
     if (isset($_POST['task_id'])) {
         $task_id = $_POST['task_id'];
         $curr_id = $_POST['curr_id'];
         $is_done = isset($_POST['is_done']) ? 1 : 0;
-        $user_id = $_POST['user_id'];
+        $user_id = $_SESSION['user_id'];
         if($is_done == 1){
             // Veritabanında görevin durumunu güncelle
             $stmt = $conn->prepare("DELETE FROM `to-do` WHERE `id` = ?");
@@ -25,7 +30,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
         }
     }
     // İşlem tamamlandıktan sonra yönlendirme yap
-    header('Location: to_do.php?id='. $curr_id .'&user_id='. $user_id);
+    header('Location: to_do.php?id='. $curr_id );
     exit();
 }
 ?>
